requirements.txt
assemblyai
streamlit==1.38.0
aiohttp
google-cloud-translate==3.18.0
google-cloud-texttospeech
fastapi
uvicorn
jinja2==3.1.4
sse_starlette
sse-starlette~=2.2.0
starlette
pydantic
python-dotenv
protobuf